gnu: llvm-for-mesa: Build all targets.
This change is needed for updating mesa to 24.2.1.  Otherwise, mesa fails to
build.  For example, linking intel_clc fails with errors of libclang-cpp.so
having undefined references to various LLVM targets.

It is possible that just some LLVM targets could be built or a modified
version of the clang input for mesa (now required) could be used to try to
reduce the closure size as before.

* gnu/packages/llvm.scm (llvm-for-mesa)[arguments]: Don't set the
LLVM_TARGETS_TO_BUILD configure flag.  Also, don't remove it from the
inherited configure-flags.  Fix indentation.

Change-Id: I4a2b7239e6ea77b0d3bce10bdf945691faa84994