REMOVEME gnu: commencement: bash-mesboot: Avoid rebuild.
XXX This avoids a boot0-world rebuild, for the gcc-14 transition we only need
to rebuild starting from gcc-boot0.

IWBN if we could program commencement in a way that it's more robust wrt base
package changes (or more flexible to counter such changes?).

Before we point the build farm to core-packages-team, we probably want to
remove this patch.

* gnu/packages/commencement.scm (bash-mesboot): Change into package definition
in order to undo #:configure-flags change made in bash-minimal.

Change-Id: I2517e7fd9e2fa3579279b2c85c4f2121ceee1a76