REMOVEME gnu: commencement: static-bash-for-glibc: Avoid rebuild.
XXX This avoids a glibc-final, gcc-final rebuild at this time.  The change to
bash-minimal should be enough for `static-bash-for-glibc' to build if we
remove this patch.

IWBN if we could program commencement in a way that it's more robust wrt base
package changes (or more flexible to counter such changes?).

Before we point the build farm to core-packages-team, we probably want to
remove this patch.

* gnu/packages/commencement.scm (REMOVEME-undo-bash-minimal-CFLAGS-change,
REMOVEME-clean-bash-configure-flags): New procedures, factored-out from...
(bash-mesboot): ...here.
* gnu/packages/commencement.scm (static-bash-for-glibc)[arguments]: Use them
to avoid a glibc-intermediate, glibc-final, gcc-final at this time.

Change-Id: I2517e7fd9e2fa3579279b2c85c4f2121ceee1a76