gnu: librewolf: Repace 'patch-config with 'expand-extension-scope.
With the Mozzarella changes moved to 'use-mozzarella, 'fix-config no longer
reflects what this phase is doing; rename it.  Also, move it to run after
'unpack, and modify the file in the source tree, rather than after 'install,
modifying the file in the package output.

* gnu/packages/librewolf.scm (librewolf):
[phase 'patch-config]: Rename to 'expand-extension-scope.  Run earlier, and
work on the source tree, not the package output.

Change-Id: If226a70daa780d652b2bb3028c888d029c765444