Revert "build: Add a commit-msg hook that embeds Change-Id in commit messages."
This reverts commit 8005e09b261d65bf0f7469cd8e89423c1c1db820 and follow-ups.
This was added 2 years ago to be able to trace commits between the git
repository and the guix-patches tracker, and to implement a solution to
automatically close merged commits (which has not materialized).

Since we have now migrated to Codeberg, the cross-referencing between reviewed
commits (pull requests) and the actual commits merged can be done in the
Codeberg UI (via the 'manually merged' option), and Codeberg also supports
closing a bug via the "Fixes: #NNN" git trailer, and as a convention we can
add a "Merges: !NNN" git trailer in a commit closing a PR (the actual closing
is not yet implemented by Forgejo), which mostly obsoletes the purpose of the
Change-Id git trailer.